Various tweaks to support building with excessive GCC warning flags
This fixes a bunch of small issues in order to enable a clean
successful build with a crazy number of GCC warning flags. A lot of
these changes are covered by -Wshadow, -Wformat-security, and
-Wstrict-overflow=5.

+++ src/pacman/util.c
@@ -258,7 +258,7 @@ void indentprint(const char *str, unsigned short indent, unsigned short cols)
 {
 	wchar_t *wcstr;
 	const wchar_t *p;
-	int len, cidx;
+	size_t len, cidx;
 
 	if(!str) {
 		return;
